fix(ethernetip): 修复重连后未继续读取的问题并修正类型使用
1. 为readGroup重连成功后补充注释说明无需返回
2. 修正protocol_verifier_test中int标签使用SetInt16而非SetInt32
3. 新增测试验证重连后继续读取的逻辑

+++ internal/driver/ethernetip/protocol_verifier_test.go
@@ -214,7 +214,8 @@ func (pv *ProtocolVerifier) verifyTagReadWrite() []TestResult {
 		fmt.Printf("✗ 单个标签写入 - 初始化失败: %v\n", err)
 		results = append(results, TestResult{Name: "单标签写_初始化", Passed: false, Message: err.Error()})
 	} else {
-		tag2.SetInt32(12345)
+		// 使用正确的数据类型：IntTag 是 INT (16位)，应该使用 SetInt16
+		tag2.SetInt16(12345)
 		err = tag2.Write()
 		if err != nil {
 			fmt.Printf("✗ 单个标签写入失败: %v\n", err)
